reorganizes iterator and view utils to remove from mesh
