+++ src/mesh/MeshMaps_decl.hh
@@ -71,6 +71,7 @@
 #pragma once
 
 #include "AmanziTypes.hh"
+#include "ViewUtils.hh"
 
 namespace Amanzi {
 namespace AmanziMesh {
